requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pdfplumber==0.11.4
pymupdf==1.24.10
python-multipart==0.0.12
pillow==10.4.0
pydantic==2.9.2
aiofiles==24.1.0
